requirements.txt
# frozen requirements generated by pip-deepfreeze
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
copier==7.0.1
commonmark==0.9.1
cryptography==38.0.1
Deprecated==1.2.13
docutils==0.19
flower==1.2.0
frozenlist==1.3.1
gidgethub==5.2.0
github3.py==3.2.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.9.3
kombu==5.2.4
lxml==4.9.1
more-itertools==8.14.0
multidict==6.0.2
OdooRPC==0.8.0
packaging==21.3
pkginfo==1.8.3
prometheus-client==0.14.1
prompt-toolkit==3.0.31
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
raven==6.10.0
readme-renderer==37.2
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
SecretStorage==3.3.3
setuptools-odoo==3.1.6
setuptools-scm==7.0.5
six==1.16.0
tomli==2.0.1
tornado==6.2
twine==4.0.1
types-cryptography==3.3.23
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
